Add ability to make IORegEntry request
